Domain Transfer: Factor out Google Domains instructions modal to component
