+++ client/landing/stepper/declarative-flow/internals/components/google-domains-transfer-instructions/style.scss
@@ -1,0 +1,25 @@
+
+
+.google-domains-transfer-instructions__modal {
+	font-size: rem(14px);
+	width: 600px;
+	max-width: 95%;
+	margin: auto auto;
+
+	summary {
+		font-weight: bold;
+		cursor: pointer;
+	}
+
+	details {
+		margin-top: 18px;
+
+		p {
+			margin: 8px 0 16px;
+		}
+	}
+}
+
+.google-domains-transfer-instructions__image {
+	border: 1px solid var(--color-neutral-10);
+}
